feat: optimized query 1

+++ api/src/main/java/ar/edu/itba/pod/query1/TotalInfractionsReducer.java
@@ -3,14 +3,14 @@ package ar.edu.itba.pod.query1;
 import com.hazelcast.mapreduce.Reducer;
 import com.hazelcast.mapreduce.ReducerFactory;
 @SuppressWarnings("deprecation")
-public class TotalInfractionsReducer implements ReducerFactory<String, Long, Long> {
+public class TotalInfractionsReducer implements ReducerFactory<String, Integer, Integer> {
     @Override
-    public Reducer<Long, Long> newReducer(String s) {
+    public Reducer<Integer, Integer> newReducer(String s) {
         return new InfractionReducer();
     }
 
-    private static class InfractionReducer extends Reducer<Long, Long>{
-        private long sum;
+    private static class InfractionReducer extends Reducer<Integer, Integer>{
+        private int sum;
 
         @Override
         public void beginReduce() {
@@ -18,12 +18,12 @@ public class TotalInfractionsReducer implements ReducerFactory<String, Long, Lon
         }
 
         @Override
-        public void reduce(Long value) {
+        public void reduce(Integer value) {
             sum += value;
         }
 
         @Override
-        public Long finalizeReduce() {
+        public Integer finalizeReduce() {
             return sum;
         }
     }
